+++ examples/reader.rs
@@ -18,7 +18,7 @@ fn main() -> Result<()> {
 
 fn start() -> Result<()> {
     let mut term = Terminal::new();
-    let mut reader = TermRead::new(&mut term, KeyCode::Esc);
+    let mut reader = TermRead::new(&mut term, KeyCode::Enter);
     reader.set_prompt("type here: ");
     println!("\n\rread: {}\r", reader.edit("old text", None)?);
     Ok(())